Final Day Commit Backend

+++ graphql/typedefs/reviewTypeDefs.js
@@ -4,7 +4,7 @@ const reviewTypeDefs = gql`
     type Review {
         reviewId: ID!
         userId: String!
-        rating: Int!
+        rating: Float!
         comment: String!
         timestamp: String!
     }
